huge refactoring WIP to split out into more modules and classes
